refactor(pins): legacy 자동캡처 스택 제거 + FrontmatterSchema 이전
