+++ lib/pins/compat.test.ts
@@ -1,6 +1,6 @@
 import { describe, test, expect } from 'bun:test';
 import { toCanonical } from './compat';
-import type { FrontmatterSchema } from '../../hooks/pin-up/types';
+import type { FrontmatterSchema } from './legacy-types';
 
 // Minimal valid legacy fixture factory
 function legacyBase(slug: string): FrontmatterSchema {
